requirements.txt
google-generativeai
gradio
python-dotenv
pandas
ics